requirements.txt
imagehash==4.3.2
Pillow==12.0.0
ExifRead==3.5.1
opencv-python==4.12.0.88
pandas==2.2.2
imageio-ffmpeg==0.6.0